requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
